Custom Gooey Keys importer (#62)
* Custom Gooey Keys importer

Heyo!

I've exhausted my patience with Google Docs, so I've adapted the work by Zekth, GreatWizard & nachie into a somewhat more generalized -- i.e. all of the variables have been pulled out of the logic -- custom importer for complete (except for the IDs) JSON catalogs.

Please let me know if there are any issues here, either in the code itself or the general concept. I'm happy to make any changes or discuss as needed.

One thing that could be an issue (specifically for my catalog) is that I've capitalized my artisan name. It is currently all lowercase on the site, but I'd like it to be in capital case.

As always, thanks for y'all's work on the archive and for your time!

* Tweaks per Zekth's input

Added a few comments and a new `ARTISAN_ID` value, for instances where the ID for a catalog is based off of a string that doesn't match the `ARTISAN_NAME` (i.e. their display name).